requirements.txt
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tensorflow==2.15.1
joblib==1.4.2
python-dotenv==1.0.1
kaggle==1.6.17






